Fixed branches and added travis support

+++ src/arch/instrs/branches.rs
@@ -2,8 +2,8 @@ use arch::cpu::CPU;
 
 pub fn bcc(cpu: &mut CPU) -> (u8, u8) {
     if !cpu.registers.getC() {
-        let addr = decode_zeropage!(cpu).0 as u16;
-        let new_pc = cpu.registers.PC.wrapping_add(addr);
+        let addr = decode_zeropage!(cpu).0 as i8;
+        let new_pc = cpu.registers.PC.wrapping_add(addr as u16);
         cpu.registers.PC = new_pc;
 
         (3, 0)
@@ -15,8 +15,8 @@ pub fn bcc(cpu: &mut CPU) -> (u8, u8) {
 
 pub fn bcs(cpu: &mut CPU) -> (u8, u8) {
     if cpu.registers.getC() {
-        let addr = decode_zeropage!(cpu).0 as u16;
-        let new_pc = cpu.registers.PC.wrapping_add(addr);
+        let addr = decode_zeropage!(cpu).0 as i8;
+        let new_pc = cpu.registers.PC.wrapping_add(addr as u16);
         cpu.registers.PC = new_pc;
 
         (3, 0)
@@ -28,8 +28,8 @@ pub fn bcs(cpu: &mut CPU) -> (u8, u8) {
 
 pub fn beq(cpu: &mut CPU) -> (u8, u8) {
     if !cpu.registers.getZ() {
-        let addr = decode_zeropage!(cpu).0 as u16;
-        let new_pc = cpu.registers.PC.wrapping_add(addr);
+        let addr = decode_zeropage!(cpu).0 as i8;
+        let new_pc = cpu.registers.PC.wrapping_add(addr as u16);
         cpu.registers.PC = new_pc;
 
         (3, 0)
@@ -41,8 +41,8 @@ pub fn beq(cpu: &mut CPU) -> (u8, u8) {
 
 pub fn bne(cpu: &mut CPU) -> (u8, u8) {
     if cpu.registers.getZ() {
-        let addr = decode_zeropage!(cpu).0 as u16;
-        let new_pc = cpu.registers.PC.wrapping_add(addr);
+        let addr = decode_zeropage!(cpu).0 as i8;
+        let new_pc = cpu.registers.PC.wrapping_add(addr as u16);
         cpu.registers.PC = new_pc;
 
         (3, 0)
@@ -54,8 +54,8 @@ pub fn bne(cpu: &mut CPU) -> (u8, u8) {
 
 pub fn bmi(cpu: &mut CPU) -> (u8, u8) {
     if !cpu.registers.getN() {
-        let addr = decode_zeropage!(cpu).0 as u16;
-        let new_pc = cpu.registers.PC.wrapping_add(addr);
+        let addr = decode_zeropage!(cpu).0 as i8;
+        let new_pc = cpu.registers.PC.wrapping_add(addr as u16);
         cpu.registers.PC = new_pc;
 
         (3, 0)
@@ -67,8 +67,8 @@ pub fn bmi(cpu: &mut CPU) -> (u8, u8) {
 
 pub fn bpl(cpu: &mut CPU) -> (u8, u8) {
     if cpu.registers.getN() {
-        let addr = decode_zeropage!(cpu).0 as u16;
-        let new_pc = cpu.registers.PC.wrapping_add(addr);
+        let addr = decode_zeropage!(cpu).0 as i8;
+        let new_pc = cpu.registers.PC.wrapping_add(addr as u16);
         cpu.registers.PC = new_pc;
 
         (3, 0)
@@ -80,8 +80,8 @@ pub fn bpl(cpu: &mut CPU) -> (u8, u8) {
 
 pub fn bvs(cpu: &mut CPU) -> (u8, u8) {
     if !cpu.registers.getV() {
-        let addr = decode_zeropage!(cpu).0 as u16;
-        let new_pc = cpu.registers.PC.wrapping_add(addr);
+        let addr = decode_zeropage!(cpu).0 as i8;
+        let new_pc = cpu.registers.PC.wrapping_add(addr as u16);
         cpu.registers.PC = new_pc;
 
         (3, 0)
@@ -93,8 +93,8 @@ pub fn bvs(cpu: &mut CPU) -> (u8, u8) {
 
 pub fn bvc(cpu: &mut CPU) -> (u8, u8) {
     if cpu.registers.getV() {
-        let addr = decode_zeropage!(cpu).0 as u16;
-        let new_pc = cpu.registers.PC.wrapping_add(addr);
+        let addr = decode_zeropage!(cpu).0 as i8;
+        let new_pc = cpu.registers.PC.wrapping_add(addr as u16);
         cpu.registers.PC = new_pc;
 
         (3, 0)
